ai_engine/requirements.txt
flask==3.1.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
scipy==1.15.2
gunicorn==23.0.0
prometheus-client==0.21.0
pytest==8.3.4
openai>=1.40.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
prophet>=1.1.5
hdbscan>=0.8.33
shap>=0.45.0
xgboost>=2.0.0
